pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nz-house-prices"
version = "1.1.1"
description = "Scrape house price estimates from NZ real estate websites"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Gio"}
]
keywords = ["house prices", "real estate", "new zealand", "scraping", "property"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "playwright>=1.40.0",
    "PyYAML>=6.0",
    "numpy>=1.26.0",
    "requests>=2.28.0",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
sheets = [
    "gspread>=6.1.0",
    "oauth2client>=4.1.3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
nz-house-prices = "nz_house_prices.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nz_house_prices = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (requires network access)",
    "live: Extended live tests (run with --live flag)",
    "performance: Performance benchmark tests (run with --performance flag)",
]
